Added some cleanup around my expenses component to allow for writing unit tests
