+++ src/app/fyle/my-expenses/my-expenses-filters.model.ts
@@ -1,0 +1,11 @@
+export type Filters = Partial<{
+  state: string[];
+  date: string;
+  customDateStart: Date;
+  customDateEnd: Date;
+  receiptsAttached: string;
+  type: string[];
+  sortParam: string;
+  sortDir: string;
+  cardNumbers: string[];
+}>;
