+++ src/app/fyle/my-expenses/my-expenses.page.html
@@ -63,7 +63,7 @@
       </ng-container>
     </ng-container>
     <ion-buttons mode="md" slot="end">
-      <ion-button *ngIf="isMergeAllowed(selectedElements)" (click)="mergeExpenses()">
+      <ion-button *ngIf="isMergeAllowed" (click)="mergeExpenses()">
         <ion-icon src="../../../assets/svg/fy-merge.svg"></ion-icon>
       </ion-button>
       <ion-button
